backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
celery==5.3.0
redis==4.5.1
gunicorn==20.1.0 